+++ src/app/components/game/allTechListSection.tsx
